Make protoc self-hosting
We convert protoc to use protojure generated code rather than flatlands
by compiling the official protoc-protobufs with protojure itself.  We are
currently hand-editing two elements of the protojure output:

- The dependency namespace (google.protobuf -> com.google.protobuf) due to [Issue 11](https://github.com/protojure/protoc-plugin/issues/11)
- We remove the :oneof-index default generated from descriptor.proto due to the lack of a "has-a" feature in protojure and the design of the protoc <-> plugin protocol requiring it.

N.B. We reduce the testing coverage from 86% to 83 due to the removal of
the 'bad-io-test', which resulted in lesser coverage through the 'abort'
handling logic in main/util.  This is due to the fact that protojure,
by design, is more forgiving of ill-formed messages, preferring instead
to optimistically return what it can interpret rather than throw an
exception.

We will devise a test to exercise the abort path in the future.

Signed-off-by: Greg Haskins <gmhaskins@statestreet.com>