Convert protobuf identifiers to idiomatic clojure
Protobuf style guide calls for snake_style and capitalization in various identifiers:

https://developers.google.com/protocol-buffers/docs/style#message-and-field-names

However, clojure style calls for kebab-style and lower-case.  This patch implements
an implicit conversion and follows the conventions already in use within flatlands:

+----------------+---------------+---------------+
|      Type      |   Protobuf    |    Clojure    |
+----------------+---------------+---------------+
| message fields | example_field | example-field |
| enum labels    | EXAMPLE_LABEL | example-label |
+----------------+---------------+---------------+

N.B. THIS IS A BREAKING CHANGE: Clients that were using earlier versions of protojure
may be referencing identifiers using the old scheme.  While inconvenient, it is ideal
to get this type of change in before we cut v1.0

Signed-off-by: Greg Haskins <gmhaskins@statestreet.com>