[#28] Support multiple .proto files with the same 'package'
We now generate output 1:1 with the number of distinct 'packages', rather
than the number of input .proto files we encounter.  The biggest caveat
is that the first instance of a package encountered becomes the authority
on things like 'java_package', and we ignore any others.  It's probably
dubious to use distinct option overrides within the same package, but
we do not validate this.

This has been tested to properly generate the well-known-types from
the Protobuf project, which is what drove the discovery of this issue
in the first place.

Fixes Issue #28

Signed-off-by: Greg Haskins <greg@manetu.com>