Support "re-export"
There is an issue with the generated code that becomes apparent for use-cases
involving "re-exporting" protobuf messages that were imported by protojure.

The issue is that subordinate messages appear in the top-level record as keys
even if the subordinate message is not present.  In this case, the subordinate
message value will be nil, which causes problems if a user tries to re-export the
top-level record.

This patch allows for the graceful handling of re-exporting protojure records
by handling the case where suborindate records are nil and skipping over them,
similar to if they were simply never set.

Signed-off-by: Greg Haskins <greg@manetu.com>